+++ src/matgl/config.py
@@ -10,6 +10,9 @@ from typing import Literal
 
 from pymatgen.core.periodic_table import Element
 
+# Coulomb conversion
+COULOMB_CONSTANT = 14.399645478425668
+
 # Default set of elements supported by universal matgl models. Excludes radioactive and most artificial elements.
 DEFAULT_ELEMENTS = tuple(el.symbol for el in Element if el.symbol not in ["Po", "At", "Rn", "Fr", "Ra"] and el.Z < 95)
 
@@ -25,16 +28,21 @@ PRETRAINED_MODELS_BASE_URL = "http://github.com/materialsvirtuallab/matgl/raw/ma
 BACKEND: Literal["PYG", "DGL"] = os.environ.get("MATGL_BACKEND", "PYG")  # type: ignore[assignment,return-value]
 
 
-if BACKEND == "DGL":
-    try:
-        importlib.util.find_spec("dgl")  # type: ignore[attr-defined]
-    except ImportError as err:
-        raise RuntimeError("Please install DGL to use this backend.") from err
-else:
-    try:
-        importlib.util.find_spec("torch_geometric")  # type: ignore[attr-defined]
-    except ImportError as err:
-        raise RuntimeError("Please install torch_geometric to use this backend.") from err
+def ensure_backend(backend: Literal["DGL", "PYG"]) -> None:
+    """Validate that the requested backend is installed."""
+    if backend == "DGL":
+        try:
+            importlib.util.find_spec("dgl")  # type: ignore[attr-defined]
+        except ImportError as err:
+            raise RuntimeError("Please install DGL to use this backend.") from err
+    else:
+        try:
+            importlib.util.find_spec("torch_geometric")  # type: ignore[attr-defined]
+        except ImportError as err:
+            raise RuntimeError("Please install torch_geometric to use this backend.") from err
+
+
+ensure_backend(BACKEND)
 
 
 def clear_cache(confirm: bool = True) -> None: