+++ tests/ext/test_ase_dgl.py
@@ -47,6 +47,20 @@ def test_PESCalculator_and_M3GNetCalculator(MoS):
     with pytest.raises(ValueError, match=r"Unsupported stress_unit: Pa. Must be 'GPa' or 'eV/A3'."):
         PESCalculator(potential=ff, stress_unit="Pa")
 
+    adaptor = AseAtomsAdaptor()
+    s_ase = adaptor.get_atoms(MoS)  # type: ignore
+    ff = load_model("pretrained_models/QET-MatQ-PES/")
+    ff.calc_hessian = True
+    calc = PESCalculator(potential=ff, state_attr=None, stress_unit="eV/A3")
+    s_ase.set_calculator(calc)
+    assert isinstance(s_ase.get_potential_energy(), float)
+    assert list(s_ase.get_forces().shape) == [2, 3]
+    assert list(s_ase.get_stress().shape) == [6]
+    assert list(s_ase.get_charges().shape) == [2]
+    assert list(calc.results["hessian"].shape) == [6, 6]
+    np.testing.assert_allclose(s_ase.get_potential_energy(), -10.876679, atol=1e-5, rtol=1e-6)
+    np.testing.assert_allclose(s_ase.get_charges(), np.array([0.727852, -0.727852]), atol=1e-5, rtol=1e-6)
+
 
 def test_CHGNetCalculator(MoS):
     adaptor = AseAtomsAdaptor()