+++ tests/layers/test_zbl_pyg.py
@@ -11,7 +11,7 @@ import matgl
 if matgl.config.BACKEND != "PYG":
     pytest.skip("Skipping PYG tests", allow_module_level=True)
 import matgl
-from matgl.layers._zbl_pyg import NuclearRepulsion
+from matgl.layers._zbl_pyg import NuclearRepulsionPyG
 
 
 @pytest.fixture
@@ -39,7 +39,7 @@ def test_nuclear_repulsion(example_data):
     element_types, graph = example_data
 
     r_cut = 3.0
-    nuclear_repulsion = NuclearRepulsion(r_cut=r_cut, trainable=True)
+    nuclear_repulsion = NuclearRepulsionPyG(r_cut=r_cut, trainable=True)
 
     energy = nuclear_repulsion(element_types, graph)