+++ changes.md
@@ -6,6 +6,17 @@ nav_order: 3
 
 # Change Log
 
+## 2.0.4
+- Bug fix for matgl.graph.data and matgl.graph.converter imports for different backends.
+
+## 2.0.3
+- Bug fix for matgl.ext.pymatgen import for different backends.
+
+## 2.0.2
+- QET (Charge-Equilibrated TensorNet) architecture and pre-trained weights are added!
+- Begun a migration to Pytorch-Geometric over the now-deprecated DGL. So far, only vanilla TensorNet has been
+  implemented in PYG). DGL models still work but require a manual setup (change of backend and installation of DGL).
+
 ## 1.2.7
 - Use original custom RemoteFile rather than fsspec, which is very finicky with SSL connections.
 - _create_directed_line_graph error handling (@bowen-bd)